backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg[binary]
huey==2.5.0
redis==5.0.1
django-cors-headers==4.3.1